pyproject.toml
[tool.poetry]
name = "operationsgateway-api"
version = "0.1.0"
description = ""
readme = "README.md"
authors = ["Matthew Richards <matthew.richards@stfc.ac.uk>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
fastapi = "^0.103.2"
motor = "^3.3.2"
uvicorn = "^0.20.0"
python-multipart = "^0.0.5"
PyYAML = "^6.0"
ImageHash = "4.3.1"
# scipy is a dependency of ImageHash but is listed to mitigate
# a vulnerability in scipy version 1.9.3
scipy = "1.10.0"
numpy = "^1.23.1"
h5py = "^3.7.0"
Pillow = "^10.0.1"
matplotlib = "^3.5.2"
gunicorn = "^20.1.0"
orjson = "3.9.10"
python-dateutil = "^2.8.2"
PyJWT = "2.4.0"
cryptography = "^41.0.0"
python-ldap = "3.4.2"
boto3 = "^1.26.97"
cron-converter = "^1.0.1"
zeep = "^4.2.1"
httpx = "^0.23.3"
# Workaround for: https://github.com/python-poetry/poetry-plugin-export/issues/183
# urllib3 isn't a dependency of this project, remove once the issue has been resolved
urllib3 = ">=1.25.4,<1.27"
lark = "1.1.9"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
flake8 = "^4.0.1"
flake8-black = "^0.2.1"
flake8-broken-line = "0.4.0"
flake8-bugbear = "^22.4.25"
flake8-builtins = "^1.5.3"
flake8-commas = "^2.1.0"
flake8-comprehensions = "^3.3.0"
flake8-import-order = "^0.18.1"
flake8-logging-format = "^0.6.0"
pep8-naming = "^0.12.1"
safety = "^2.3.1"
pytest = "^7.2.0"
requests = "2.31.0"
pytest-cov = "^3.0.0"
coverage = "^6.4.4"
pytest-asyncio = "^0.21.0"
s4cmd = "^2.1.0"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.28.0"

[tool.coverage.paths]
source = ["operationsgateway_api"]

[tool.coverage.run]
branch = true
source = ["operationsgateway_api"]

[tool.coverage.report]
show_missing = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
